refactor FileType
